add Spring security

+++ src/main/java/com/example/demo/controller/MemberRestController.java
@@ -4,6 +4,9 @@ import com.example.demo.entity.*;
 import com.example.demo.services.IMemberService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -33,32 +36,53 @@ public class MemberRestController {
     }
 
     @GetMapping(value = "/{id}")
-    public MemberReturn findMembersById(@PathVariable("id") String id) {
-        MemberReturn mr = new MemberReturn();
-        BeanUtils.copyProperties( memberService.findDistinctByPublicID(id),mr);
-        return mr;
+    public MemberReturn findMembersById(@PathVariable("id") String publicId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            if (memberService.findByEmail(currentUserName).getPublicID().equals(publicId) || memberService.findByEmail(currentUserName).hasRole("ADMIN")) {
+                MemberReturn mr = new MemberReturn();
+                BeanUtils.copyProperties(memberService.findDistinctByPublicID(publicId), mr);
+                return mr;
+            }
+        }
+        return null;
     }
+
 
     @PostMapping(value = "/etudiant")
     public MemberReturn addEtudiant(@RequestBody Etudiant e) {
         MemberReturn mr = new EtudiantReturn();
-        Member m=memberService.addMember(e);
+        Member m = memberService.addMember(e);
         BeanUtils.copyProperties(m, mr);
         return mr;
     }
 
     @DeleteMapping(value = "/{publicId}")
-    public void deleteEtudiant(@PathVariable("id") String publicId) {
-        memberService.deleteMember(publicId);
+    public void deleteEtudiant(@PathVariable("publicId") String publicId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            if (memberService.findByEmail(currentUserName).getPublicID().equals(publicId) || memberService.findByEmail(currentUserName).hasRole("ADMIN"))
+                memberService.deleteMember(publicId);
+        }
+
     }
 
     @PutMapping(value = "/etudiant/{publicId}")
     public MemberReturn updateEtudiant(@PathVariable("publicId") String publicId, @RequestBody Etudiant e) {
-        Long id = memberService.findDistinctByPublicID(publicId).getId();
-        e.setId(id);
-        MemberReturn mr = new EtudiantReturn();
-        BeanUtils.copyProperties(memberService.updateMember(e), mr);
-        return mr;
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            if (memberService.findByEmail(currentUserName).getPublicID().equals(publicId) || memberService.findByEmail(currentUserName).hasRole("ADMIN")) {
+                Long id = memberService.findDistinctByPublicID(publicId).getId();
+                e.setId(id);
+                MemberReturn mr = new EtudiantReturn();
+                BeanUtils.copyProperties(memberService.updateMember(e), mr);
+                return mr;
+            }
+        }
+        return null;
     }
 
     @PostMapping(value = "/enseignant")
@@ -69,12 +93,20 @@ public class MemberRestController {
     }
 
     @PutMapping(value = "/enseignant/{id}")
-    public MemberReturn updateEnseignant(@PathVariable("id") String pubId, @RequestBody EnseignantChercheur e) {
-        Long id = memberService.findDistinctByPublicID(pubId).getId();
-        e.setId(id);
-        MemberReturn mr = new EnseignantChercheurReturn();
-        BeanUtils.copyProperties(memberService.updateMember(e), mr);
-        return mr;
+    public MemberReturn updateEnseignant(@PathVariable("id") String publicId, @RequestBody EnseignantChercheur e) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            if (memberService.findByEmail(currentUserName).getPublicID().equals(publicId) || memberService.findByEmail(currentUserName).hasRole("ADMIN")) {
+
+                Long id = memberService.findDistinctByPublicID(publicId).getId();
+                e.setId(id);
+                MemberReturn mr = new EnseignantChercheurReturn();
+                BeanUtils.copyProperties(memberService.updateMember(e), mr);
+                return mr;
+            }
+        }
+        return null;
     }
 
 }
